created new model for app directory, file names
